+++ CHANGELOG.md
@@ -11,6 +11,7 @@ We go to the next version after each release on the
 
 - Backlight Toggle Button Visibility in Quick Settings.
 - Do Not Disturb Toggle Button Visibility in Quick Settings by @michael.online.
+- GNOME Shell 50 Support.
 
 ### Fixed